ParticlePDBVisualizer and MeshPDBVisualizer inherit from abstract PDBVisualizer, ready to implement atoms for MeshPDBVisualizer
