+++ Assets/Molecules/Palettes/MoleculePalette.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4d0e856ee271f4df1a73f6c08df80029
-timeCreated: 1481945171
+guid: 61b5859bdb1be4f15b9e42c41312dedc
+timeCreated: 1482027638
 licenseType: Free
 NativeFormatImporter:
   userData:
