+++ Assets/Scripts/Visualization/PDBVisualizer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 79ecef9662fd541baa2be1dbc0bb8a61
+timeCreated: 1482025701
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
